requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
edge-tts==7.0.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.1
multidict==6.1.0
packaging==24.2
propcache==0.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.12.2
Werkzeug==3.1.3
yarl==1.18.3
